Fix .link-text appearance
Fixing .link-text class to make the text appear gradually together to the navbar by the use of opacity, visibility and transition for such properties.

I switched to display: inline, it was pointless to set display: none by default since the text wouldn't appear no matter what in large screens given the fact that the '.navbar when not :hover' wouldn't allow it to be seen with its 5rem width.

I added the 'display: none' property to .link-text while in small screens to avoid the menu layout to be broken by such changes.